removed regexp title matching for relation import
