+++ app/api/csv/typeParsers.js
@@ -1,18 +1,15 @@
-import entities from 'api/entities';
 import url from 'url';
-
-import { unique, emptyString } from 'api/utils/filters';
 
 import geolocation from './typeParsers/geolocation.js';
 import multiselect from './typeParsers/multiselect.js';
 import select from './typeParsers/select.js';
+import relationship from './typeParsers/relationship.js';
 
 export default {
   geolocation,
-
   select,
-
   multiselect,
+  relationship,
 
   async default(entityToImport, templateProperty) {
     return this.text(entityToImport, templateProperty);
@@ -42,34 +39,5 @@ export default {
       label,
       url: linkUrl,
     };
-  },
-
-  async relationship(entityToImport, templateProperty) {
-    const values = entityToImport[templateProperty.name].split('|')
-    .map(v => v.trim())
-    .filter(emptyString)
-    .filter(unique);
-
-    const current = await entities.get({ title: { $in: values.map(v => new RegExp(`\\s?${v}\\s?`, 'i')) } });
-    const newValues = values.filter(v => !current.map(c => c.title.trim()).includes(v));
-
-    await newValues.reduce(
-      async (promise, title) => {
-        await promise;
-        return entities.save({
-          title,
-          template: templateProperty.content,
-        },
-          {
-            language: 'en',
-            user: {}
-          }
-        );
-      },
-      Promise.resolve([])
-    );
-
-    const toRelateEntities = await entities.get({ title: { $in: values.map(v => new RegExp(`\\s?${v}\\s?`, 'i')) } });
-    return toRelateEntities.map(e => e.sharedId).filter((id, index, ids) => ids.indexOf(id) === index);
   },
 };
